requirements.txt
gradio>=4.0.0
python-dotenv>=1.0.0
mailjet-rest>=1.3.0
pydantic>=2.0.0
openai>=1.102.0
openai-agents>=0.2.10
